Simplify JS example by extracting StackClientApp configuration
