+++ examples/js-example/credential.html
@@ -34,12 +34,6 @@
     </div>
   </div>
 
-  <div id="userInfo" class="hidden">
-    <h2>User Information</h2>
-    <p>Email: <span id="userEmail"></span></p>
-    <button id="signOut">Sign Out</button>
-  </div>
-
   <script type="module" src="/credential-script.ts"></script>
 </body>
 </html>
