+++ examples/js-example/oauth.html
@@ -4,11 +4,6 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Stack Auth OAuth Example</title>
-  <style>
-    .hidden {
-      display: none;
-    }
-  </style>
 </head>
 <body>
   <h1>Stack Auth OAuth Example</h1>
@@ -17,12 +12,6 @@
   <div id="loginButtons">
     <h2>Sign In with OAuth</h2>
     <button id="googleSignIn">Sign in with Google</button>
-  </div>
-
-  <div id="userInfo" class="hidden">
-    <h2>User Information</h2>
-    <p>Email: <span id="userEmail"></span></p>
-    <button id="signOut">Sign Out</button>
   </div>
 
   <script type="module" src="/oauth-script.ts"></script>
